% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-method.R
\name{auto_plot}
\alias{auto_plot}
\title{Auto plot cohort data frame}
\usage{
auto_plot(data, type)
}
\arguments{
\item{data}{dataframe with class cohort_df}

\item{type}{type of plot}
}
\description{
This auto plot function use ggplot2 as back end
}
