% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{cohort}
\alias{cohort}
\title{Create cohort data}
\usage{
cohort(
  data,
  id_col,
  time_col,
  time_unit = c("week", "month", "quarter", "year"),
  percent_form = TRUE,
  relative_time = TRUE,
  all_group = FALSE
)
}
\arguments{
\item{data}{Input data}

\item{id_col}{Variable to identify entities (maybe customer ID)}

\item{time_col}{Time with respect to id_col}

\item{time_unit}{Time unit to group to cohort}

\item{percent_form}{Should return table contain percentage value or absolute value}

\item{relative_time}{Should return column is relative time or absolute period}

\item{all_group}{Should result include group of all entities. This params
must only set to TRUE when both percent_form and relative_time is TRUE}
}
\value{
a data frame with class cohort_df
}
\description{
Perform cohort analysis on input data frame. Which is group entities who share the same period of time
and there appearance in the next period
}
