% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{cohort_count}
\alias{cohort_count}
\title{Create cohort data}
\usage{
cohort_count(
  data,
  .id_col,
  .time_col,
  time_unit = c("week", "month", "quarter", "year"),
  percent_form = TRUE,
  relative_time = TRUE,
  all_group = FALSE
)
}
\arguments{
\item{data}{Input data}

\item{.id_col}{Variable to identify entities (maybe customer ID)}

\item{.time_col}{Time correspond to id_col. Must be in date or date-time format}

\item{time_unit}{Time unit to group to cohort}

\item{percent_form}{Should return table contain percentage value or absolute value}

\item{relative_time}{Should return column is relative time or absolute period}

\item{all_group}{Should result include group of all entities. This argument
must only set to TRUE when both percent_form and relative_time is TRUE}
}
\value{
a data frame with class cohort_df
}
\description{
cohort_count() perform cohort analysis by counting the appearance of \code{id_col}. \code{id_col} will only count one time for all appearance in each time period.
This cohort analysis is suitable to measure customer retention. To summarize for example revenue of each cohort, use \code{cohort_summarise()}
}
