% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort.R
\name{cohort_summarise}
\alias{cohort_summarise}
\title{Create cohort data}
\usage{
cohort_summarise(
  data,
  .id_col,
  .time_col,
  .summarise_col,
  .fn = c("mean"),
  ...,
  time_unit = c("week", "month", "quarter", "year"),
  percent_form = TRUE,
  relative_time = TRUE,
  all_group = FALSE
)
}
\arguments{
\item{data}{Input data}

\item{.id_col}{Variable to identify entities (maybe customer ID)}

\item{.time_col}{Time correspond to id_col. Must be in date or date-time format}

\item{.summarise_col}{Variable to perform summarise on}

\item{.fn}{Function to perform summarise, in character form.default is "\code{sum}". This
function should return length 1 numeric value,}

\item{...}{Additional argument pass to \code{.fn}}

\item{time_unit}{Time unit to group to cohort}

\item{percent_form}{Should return table contain percentage value or absolute value}

\item{relative_time}{Should return column is relative time or absolute period}

\item{all_group}{Should result include group of all entities. This argument
must only set to TRUE when both percent_form and relative_time is TRUE}
}
\description{
cohort_summarise() perform cohort analysis by summarizing variable. For example, sum() of revenue of each cohort.
}
