% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfm.R
\name{rfm_score}
\alias{rfm_score}
\title{Compute RFM metrics}
\usage{
rfm_score(data, .id_col, .time_col, .m_col = NULL, cur_time = Sys.time())
}
\arguments{
\item{data}{Input data}

\item{.id_col}{Variable to identify entities (maybe customer ID)}

\item{.time_col}{Time correspond to id_col. Must be in date or date-time format}

\item{.m_col}{Column to compute monetary score, default to not compute}
}
\description{
Recently, Frequentcy, Monetary score
}
